// src/shared/components/toast/enum.ts
export enum Type {
    Success = 'success',
    Alert = 'alert',
    Warning = 'warning',
    Info = 'info',
    Error = "Error"
}
export enum Appearance {
    Light = 'light',
    Dark = 'dark',
    Glass = 'glass',
}
export enum Position {
    TopLeft = 'top-left',
    TopRight = 'top-right',
    BottomRight = 'bottom-right',
    BottomLeft = 'bottom-left',
}
export enum Layout {
    Left = 'left',
    Right = 'right',
    Top = 'top',
    Bottom = 'bottom',
}